bot/requirements.txt
discord.py
pytz~=2021.1
mysql-connector-python~=8.0.23
mysql
xlsxwriter